factory-test: Replace xinit & xterm with mutter & gnome-terminal
Current eos-factory-test.service executes the command with xinit & term.
However, EOS 7 is going to base on GNOME OS, which does not include
xinit & xterm. Then, factory-test should use mutter as the display
server in wayland mode and gnome-terminal instead.

The eos-factory-test.service needs:
* Make root user login to get the login session and running directory
  for mutter in wayland mode with PAMName [1]. When an user logins, the
  pam_systemd registers user sessions with the systemd login manager
  systemd-logind.service. It prepares environment "XDG_RUNTIME_DIR" and
  "XDG_SESSION_ID" for mutter in wayland mode.
* Give the user session a seat by setting TTYPath=/dev/tty1 [2]. Then,
  mutter can show up to front.

And, because the service will execute mutter and gnome-terminal, instead
of xinit and xterm, change the package dependencies accrodingly.

[1]: https://www.freedesktop.org/software/systemd/man/latest/systemd.exec.html#PAMName=
[2]: https://www.freedesktop.org/software/systemd/man/latest/systemd.exec.html#TTYPath=

Fixes: https://github.com/endlessm/eos-build-meta/issues/65